b/316247020 Rename Result class
Use AnnotatedResult as a less generic name for the class
